a2a_multi_agent/requirements.txt
a2a-sdk==0.3.0
langchain==0.3.27
langgraph==0.6.2
google-genai==1.28.0
httpx==0.28.1
python-dotenv==1.1.1
langchain-google-genai==2.1.8
uvicorn==0.35.0
click==8.2.1
rich==14.1.0
tzdata
streamlit